+++ CHANGELOG.md
@@ -1,10 +1,13 @@
 ## 2.0.0
-- Upgraded dependencies to PSR-17
+- Migrated HttpClient to PSR-17 and PSR-18
+- Updated README
 
 **Breaking Changes:**
 - Changed type of `$errors` in `PaynowException` to `Error`
 - Changed the name of method `getErrorType` for `Error`
-- Changed type of `$data` to string for `SignatureCalculator` 
+- Changed type of `$data` to string for `SignatureCalculator`
+- Changed `Payment::authorize` response to `Authorize`
+- Changed `Payment::status` response to `Status`
 
 ## 1.0.6
 - Marked `getErrorType` for `Error` as deprecated